Rename utility and documentation files
