+++ .gitignore
@@ -1,12 +1,3 @@
-__pycache__/all_routes.cpython-313.pyc
-__pycache__/config.cpython-313.pyc
-__pycache__/functions.cpython-313.pyc
-routes/__pycache__/__init__.cpython-313.pyc
-routes/__pycache__/api.cpython-313.pyc
-routes/__pycache__/auth.cpython-313.pyc
-routes/__pycache__/collab.cpython-313.pyc
-routes/__pycache__/diagrams.cpython-313.pyc
-routes/__pycache__/main.cpython-313.pyc
-routes/__pycache__/upload.cpython-313.pyc
+__pycache__/
 .idea/workspace.xml
 .DS_Store
