+++ structure.txt
@@ -25,7 +25,7 @@
 │       └── output_ProcessCOB.json
 ├── error.log
 ├── extensions.py
-├── functions.py
+├── utils.py
 ├── instance
 ├── logs
 │   ├── rules_central.log
@@ -78,7 +78,7 @@
 │   │   │   ├── loadDiagram.js
 │   │   │   └── zoom.js
 │   │   ├── activity-monitor.js
-│   │   ├── activityLogger.js
+│   │   ├── activity-logger.js
 │   │   ├── api_test_utility.js
 │   │   ├── catalog.bak
 │   │   ├── catalog.js
@@ -114,8 +114,7 @@
 │   ├── catalog.html
 │   ├── config.html
 │   ├── diagram_viewer.html
-│   ├── full_help.html
-│   ├── full_help.html~
+│   ├── documentation.html
 │   ├── help_system.html
 │   ├── hierarchy_viewer.html
 │   ├── includes
